refactor: rename record/runtime to record/dynamic
