+++ src/record/mod.rs
@@ -1,17 +1,17 @@
+pub mod dynamic;
 pub mod key;
 pub mod option;
-pub mod runtime;
 #[cfg(test)]
 pub(crate) mod test;
 
 use std::{error::Error, fmt::Debug, io, sync::Arc};
 
 use arrow::{array::RecordBatch, datatypes::Schema as ArrowSchema};
+pub use dynamic::*;
 use fusio_log::{Decode, Encode};
 pub use key::*;
 use option::OptionRecordRef;
 use parquet::{arrow::ProjectionMask, format::SortingColumn, schema::types::ColumnPath};
-pub use runtime::*;
 use thiserror::Error;
 
 use crate::inmem::immutable::ArrowArrays;
